+++ CHANGELOG.rst
@@ -18,11 +18,17 @@ Enhancements:
 - Annotations can be selected inside ``Queryset.values()`` and ``Queryset.values_list()`` expressions.
 - Added support for Python 3.8
 - The Foreign Key property is now ``await``-able as long as one didn't populate it via ``.prefetch_related()``
+- One can now specify compound indexes in the ``Meta:`` class using ``indexes``. It works just like ``unique_toghether``.
+- ``unique_together`` and ``indexes`` will correctly map the foreign key if referenced by name.
 
 Bugfixes:
 ^^^^^^^^^
 - The generated index name now has significantly lower chance of collision.
 - The compiled SQL query contains HAVING and GROUP BY only for aggregation functions.
+- Fields for FK relations are quoted properly.
+- Fields are quoted properly in ``UNIQUE`` statements.
+- Fields are quoted properly in ``KEY`` statements.
+- Comment Fields are quoted properly in PostgreSQL dialect.
 
 Deprecations:
 ^^^^^^^^^^^^^