+++ CHANGELOG.rst
@@ -19,7 +19,6 @@ Enhancements:
 - Added support for Python 3.8
 - The Foreign Key property is now ``await``-able as long as one didn't populate it via ``.prefetch_related()``
 - One can now specify compound indexes in the ``Meta:`` class using ``indexes``. It works just like ``unique_toghether``.
-- ``unique_together`` and ``indexes`` will correctly map the foreign key if referenced by name.
 
 Bugfixes:
 ^^^^^^^^^
@@ -29,6 +28,7 @@ Bugfixes:
 - Fields are quoted properly in ``UNIQUE`` statements.
 - Fields are quoted properly in ``KEY`` statements.
 - Comment Fields are quoted properly in PostgreSQL dialect.
+- ``unique_together`` and ``indexes`` will correctly map the foreign key if referenced by name.
 
 Deprecations:
 ^^^^^^^^^^^^^